move 'design' page only toolbar items into DesignPage
