+++ Composer/packages/client/src/components/ToolBar/ToolBar.tsx
@@ -3,13 +3,10 @@
 
 /** @jsx jsx */
 import { jsx } from '@emotion/core';
-import { useCallback, Fragment, useMemo } from 'react';
+import { useCallback, Fragment } from 'react';
 import formatMessage from 'format-message';
 import { ActionButton, CommandButton } from 'office-ui-fabric-react/lib/Button';
-import { DialogInfo } from '@bfc/shared';
-import get from 'lodash/get';
 
-import { VisualEditorAPI } from '../../pages/design/FrameAPI';
 import { useStoreContext } from '../../hooks/useStoreContext';
 
 import { IToolBarItem } from './ToolBar.types';
@@ -17,10 +14,6 @@ import { actionButton, leftActions, rightActions, headerSub } from './ToolBarSty
 
 type ToolbarProps = {
   toolbarItems?: Array<IToolBarItem>;
-  currentDialog?: DialogInfo;
-  onCreateDialogComplete?: (...args: any[]) => void;
-  openNewTriggerModal?: () => void;
-  showSkillManifestModal?: () => void;
 };
 
 function itemList(item: IToolBarItem, index: number) {
@@ -59,29 +52,10 @@ function itemList(item: IToolBarItem, index: number) {
 // action = {type:action/element, text, align, element, buttonProps: use
 // fabric-ui IButtonProps interface}
 export function ToolBar(props: ToolbarProps) {
+  const { toolbarItems = [], ...rest } = props;
   const {
-    toolbarItems = [],
-    currentDialog,
-    onCreateDialogComplete,
-    openNewTriggerModal,
-    showSkillManifestModal,
-    ...rest
-  } = props;
-  const {
-    actions: { onboardingAddCoachMarkRef, createDialogBegin, exportToZip },
-    state: { projectId, visualEditorSelection },
+    actions: { onboardingAddCoachMarkRef },
   } = useStoreContext();
-
-  const { actionSelected, showDisableBtn, showEnableBtn } = useMemo(() => {
-    const actionSelected = Array.isArray(visualEditorSelection) && visualEditorSelection.length > 0;
-    if (!actionSelected) {
-      return {};
-    }
-    const selectedActions = visualEditorSelection.map((id) => get(currentDialog?.content, id));
-    const showDisableBtn = selectedActions.some((x) => get(x, 'disabled') !== true);
-    const showEnableBtn = selectedActions.some((x) => get(x, 'disabled') === true);
-    return { actionSelected, showDisableBtn, showEnableBtn };
-  }, [visualEditorSelection]);
 
   const left: IToolBarItem[] = [];
   const right: IToolBarItem[] = [];
@@ -103,92 +77,8 @@ export function ToolBar(props: ToolbarProps) {
   return (
     <div aria-label={formatMessage('toolbar')} css={headerSub} role="region" {...rest}>
       <div css={leftActions}>
-        {window.location.href.includes('/dialogs/') && (
-          <div ref={addNewRef}>
-            <CommandButton
-              css={actionButton}
-              data-testid="AddFlyout"
-              iconProps={{ iconName: 'Add' }}
-              menuProps={{
-                items: [
-                  {
-                    'data-testid': 'FlyoutNewDialog',
-                    key: 'adddialog',
-                    text: formatMessage('Add new dialog'),
-                    onClick: () => {
-                      createDialogBegin([], onCreateDialogComplete);
-                    },
-                  },
-                  {
-                    'data-testid': 'FlyoutNewTrigger',
-                    key: 'addtrigger',
-                    text: formatMessage(`Add new trigger on {displayName}`, {
-                      displayName: currentDialog?.displayName ?? '',
-                    }),
-                    onClick: () => {
-                      openNewTriggerModal?.();
-                    },
-                  },
-                ],
-              }}
-              text={formatMessage('Add')}
-            />
-          </div>
-        )}
+        {window.location.href.includes('/dialogs/') && <div ref={addNewRef}></div>}
         {left.map(itemList)}{' '}
-        {window.location.href.includes('/dialogs/') && (
-          <CommandButton
-            css={actionButton}
-            disabled={!actionSelected}
-            iconProps={{ iconName: 'RemoveOccurrence' }}
-            menuProps={{
-              items: [
-                {
-                  key: 'disable',
-                  text: formatMessage('Disable'),
-                  disabled: !showDisableBtn,
-                  onClick: () => {
-                    VisualEditorAPI.disableSelection();
-                  },
-                },
-                {
-                  key: 'enable',
-                  text: formatMessage('Enable'),
-                  disabled: !showEnableBtn,
-                  onClick: () => {
-                    VisualEditorAPI.enableSelection();
-                  },
-                },
-              ],
-            }}
-            text={formatMessage('Disable')}
-          />
-        )}
-        {window.location.href.includes('/dialogs/') && (
-          <CommandButton
-            css={actionButton}
-            iconProps={{ iconName: 'OpenInNewWindow' }}
-            menuProps={{
-              items: [
-                {
-                  key: 'zipexport',
-                  text: formatMessage('Export assets to .zip'),
-                  onClick: () => {
-                    exportToZip({ projectId });
-                  },
-                },
-                {
-                  key: 'exportAsSkill',
-                  text: formatMessage('Export as skill'),
-                  onClick: () => {
-                    showSkillManifestModal?.();
-                  },
-                },
-              ],
-            }}
-            text={formatMessage('Export')}
-          />
-        )}
       </div>
       <div css={rightActions}>{right.map(itemList)}</div>
     </div>
